# 8357531: The `SegmentBulkOperations::fill` method can be improved using overlaps