This PR builds on a concept John Rose told me about some time ago. Instead of combining memory operations of various sizes, a single large and skewed memory operation can be made to clean up the tail of remaining bytes.

This has the effect of simplifying and shortening the code while improving performance. The number of branches to evaluate is reduced.