@wenshao 2025-05-22
Here you can use putShort/putInt/putLong to replace putShortUnaligned/putIntUnaligned/putLongUnaligned