@minborg 2025-05-22
> > Performance on an M1 Mac (Sequoia 15.4.1)
> 
> Are the "regressions" for very big sizes related to this patch?

By default, only segments that are smaller than 32 bytes will use the affected code.