@minborg 2025-05-22
> Here you can use putShort/putInt/putLong to replace putShortUnaligned/putIntUnaligned/putLongUnaligned

The segment might be unaligned. For example, if it was sliced off an aligned segment.