@rose00 2025-05-22
Related discussion at the hardware level:

https://github.com/openjdk/jdk/pull/25147#issuecomment-2902463076

I think I want to look into the VM, see if it can use overlapping reads and writes as well.