@minborg 2025-05-27
> > Can we remove all meaningless `!Architecture.isLittleEndian()` calls and use the platform-specific endianness unsafe primitives instead? For compensation, we can add a check in each of fill2/fill3/fill4 that `assert value == Integer/Short/Long.reverseBytes(value)` if you think that is necessary.
> 
> Let's take a look at that under a separate PR.

https://bugs.openjdk.org/browse/JDK-8357857