@minborg 2025-08-05
The new performance improvements gave some 8% better performance for smaller segments.

Before improvements

```
Benchmark                              (ELEM_SIZE)  Mode  Cnt  Score   Error  Units
SegmentBulkFill.nativeSegmentFillJava            2  avgt   30  1.322 ± 0.020  ns/op
SegmentBulkFill.nativeSegmentFillJava            3  avgt   30  1.313 ± 0.009  ns/op
SegmentBulkFill.nativeSegmentFillJava            4  avgt   30  1.323 ± 0.023  ns/op
SegmentBulkFill.nativeSegmentFillJava            5  avgt   30  1.309 ± 0.006  ns/op
SegmentBulkFill.nativeSegmentFillJava            6  avgt   30  1.310 ± 0.017  ns/op
SegmentBulkFill.nativeSegmentFillJava            7  avgt   30  1.308 ± 0.004  ns/op
SegmentBulkFill.nativeSegmentFillJava            8  avgt   30  1.312 ± 0.008  ns/op
SegmentBulkFill.nativeSegmentFillJava           12  avgt   30  1.316 ± 0.025  ns/op
```

After improvements

```
Benchmark                              (ELEM_SIZE)  Mode  Cnt  Score   Error  Units
SegmentBulkFill.nativeSegmentFillJava            2  avgt   30  1.230 ± 0.034  ns/op
SegmentBulkFill.nativeSegmentFillJava            3  avgt   30  1.228 ± 0.037  ns/op
SegmentBulkFill.nativeSegmentFillJava            4  avgt   30  1.243 ± 0.034  ns/op
SegmentBulkFill.nativeSegmentFillJava            5  avgt   30  1.230 ± 0.026  ns/op
SegmentBulkFill.nativeSegmentFillJava            6  avgt   30  1.224 ± 0.025  ns/op
SegmentBulkFill.nativeSegmentFillJava            7  avgt   30  1.232 ± 0.039  ns/op
SegmentBulkFill.nativeSegmentFillJava            8  avgt   30  1.208 ± 0.009  ns/op
SegmentBulkFill.nativeSegmentFillJava           12  avgt   30  1.242 ± 0.030  ns/op
SegmentBulkFill.nativeSegmentFillJava           24  avgt   30  1.932 ± 0.067  ns/o
```